pluggramd/requirements.txt
zmq
tinyrpc
msgpack
pillow